requirements.txt
Pillow
numpy
pytesseract
requests
opencv-python
google-cloud-datastore